+++ WoueShop.Shared/Abstraction/Error.cs
@@ -1,10 +1,47 @@
-﻿namespace WoueShop.Shared.Abstraction
+﻿namespace WoueShop.Shared.Abstraction;
+
+///<summary>
+///     Custom defined Error class for exception handling
+///</summary>
+public sealed record Error
 {
-    ///<summary>
-    ///     Custom defined Error class for exception handling
-    ///</summary>
-    public sealed record Error(string code, string? description = null)
+    public static readonly Error None = new(string.Empty, string.Empty, ErrorType.Failure);
+
+    public static readonly Error NullValue = new("Null Value", "Null value was provided", ErrorType.Failure);
+
+    public static implicit operator Result(Error error) => Result.Failure(error);
+
+    public string Code { get; }
+
+    public string Description { get; }
+
+    public ErrorType Type { get; }
+
+    private Error(string code, string description, ErrorType errorType)
     {
-        public static readonly Error None = new(string.Empty);
+        Code = code;
+        Description = description;
+        Type = errorType;
     }
+
+    public static Error NotFound(string code, string description) =>
+        new(code, description, ErrorType.NotFound);
+
+    public static Error Validation(string code, string description) =>
+        new(code, description, ErrorType.NotFound);
+
+    public static Error Failure(string code, string description) =>
+        new(code, description, ErrorType.NotFound);
+
+    public static Error Conflict(string code, string description) =>
+        new(code, description, ErrorType.NotFound);
+
+}
+
+public enum ErrorType
+{
+    Failure = 0,
+    Validation = 1,
+    NotFound = 2,
+    Conflict = 3
 }