+++ WoueShop.Shared/Abstraction/Result.cs
@@ -1,30 +1,54 @@
-﻿namespace WoueShop.Shared.Abstraction
+﻿using System.Net.Http.Headers;
+using System.Numerics;
+
+namespace WoueShop.Shared.Abstraction;
+
+/// <summary>
+/// To handle exception errors and failure for Result pattern
+/// </summary>
+public  class Result
 {
-    /// <summary>
-    /// To handle exception errors and failure for Result pattern
-    /// </summary>
-    public sealed class Result
+    public bool IsSuccess { get; }
+
+    public bool IsFailure => !IsSuccess;
+
+    public Error Error { get; }
+
+    protected internal Result(bool isSuccess, Error error)
     {
-        public bool IsSuccess { get; }
-
-        public bool IsFailure => !IsSuccess;
-
-        public Error Error { get; }
-
-        private Result(bool isSuccess, Error error)
+        if (isSuccess && error != Error.None ||
+            !isSuccess && error == Error.None)
         {
-            if (isSuccess && error != Error.None ||
-                !isSuccess && error == Error.None)
-            {
-                throw new ArgumentException("Invalid error", nameof(error));
-            }
-
-            IsSuccess = IsSuccess;
-            Error = error;
+            throw new ArgumentException("Invalid error", nameof(error));
         }
 
-        public static Result Success() => new(true, Error.None);
-
-        public static Result Failure(Error error) => new(false, error);
+        IsSuccess = IsSuccess;
+        Error = error;
     }
+
+    public static Result Success() => new(true, Error.None);
+
+    public static Result Failure(Error error) => new(false, error);
+
+    public static Result<TValue> Success<TValue>(TValue value) => new(value, true, Error.None);
+
+    public static Result<TValue> Failure<TValue>(Error error) => new(default, false, error);
+}
+
+public class Result<TValue> : Result
+{
+    private readonly TValue? _value;
+
+    protected internal Result(TValue? value, bool isSuccess, Error error)
+        : base(isSuccess, error)
+    {
+        _value = value;
+    }
+
+    public TValue? Value => IsSuccess ?
+        _value :
+        throw new InvalidOperationException("The value of a failure result could not be accessed");
+
+    public static implicit operator Result<TValue>(TValue? value) =>
+        value is not null ? Success(value) : Failure<TValue>(Error.NullValue);
 }