+++ WoueShop.Shared/Globals/UserConstants.cs
@@ -1,0 +1,14 @@
+﻿namespace WoueShop.Shared.Globals
+{
+    public static class UserConstants
+    {
+        public static readonly string AdminRole = "SuperAdmin";
+
+        public static readonly string AdminEmail = "superadmin@woue.com";
+
+        public static readonly string AdminPassword = "Developer#1";
+
+        public static readonly string CustomerRole = "Customer";
+
+    }
+}