+++ WoueShop.Shared/ReturnModels/UserInfoModel.cs
@@ -8,8 +8,6 @@
 
         public string Email { get; set; }
 
-        public string? Username { get; set; }
-
         public string Password { get; set; }
 
     }